requirements.txt
asgiref==3.2.7
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
blessed==1.17.5
boto3==1.12.48
botocore==1.15.48
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==2.2.9
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
htmlmin==0.1.12
idna==2.7
importlib-metadata==1.6.0
jmespath==0.9.5
jsonschema==3.2.0
mysqlclient==1.4.6
paramiko==2.7.1
pathspec==0.5.9
pycparser==2.20
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.0
python-memcached==1.59
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
s3transfer==0.3.3
semantic-version==2.5.0
sentry-sdk==0.14.3
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0
